added confirmation before user is deleted

+++ immigration-analytics/src/main/java/com/nineplusten/app/view/AccountManagementController.java
@@ -20,6 +20,8 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.control.Label;
 import javafx.scene.control.Button;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 
 public class AccountManagementController {
 
@@ -38,25 +40,32 @@ public class AccountManagementController {
 
   DeleteUserService deleteService;
 
-
   @FXML
-  private void refreshPage(ActionEvent event){
+  private void refreshPage(ActionEvent event) {
     userService.restart();
   }
 
   @FXML
-  private void deleteUser(ActionEvent event){
-
+  private void deleteUser(ActionEvent event) {
     String userID = userTable.getSelectionModel().getSelectedItem().userIDProperty().getValue();
 
-    deleteService = new DeleteUserService(userID);
-    deleteService.start();
-    deleteService.setOnSucceeded(new EventHandler<WorkerStateEvent>(){
-      @Override
-      public void handle(WorkerStateEvent e){
-        userService.restart();
-      }
-    });
+    Alert alert = new Alert(AlertType.CONFIRMATION);
+    alert.setTitle("Confirm");
+    alert.setHeaderText("Deleting User");
+    alert.setContentText("Are you sure you want to delete the user: " + userID + "?");
+
+    alert.showAndWait();
+
+    if (alert.getResult().getText().equals("OK")) {
+      deleteService = new DeleteUserService(userID);
+      deleteService.start();
+      deleteService.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
+        @Override
+        public void handle(WorkerStateEvent e) {
+          userService.restart();
+        }
+      });
+    }
   }
 
   @FXML
@@ -71,11 +80,11 @@ public class AccountManagementController {
     userTable.disableProperty().bind(userService.runningProperty());
 
     userService.start();
-    userService.setOnSucceeded(new EventHandler<WorkerStateEvent>(){
-      @Override 
-      public void handle(WorkerStateEvent e) { 
+    userService.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
+      @Override
+      public void handle(WorkerStateEvent e) {
         userTable.setItems(parseUsers());
-      } 
+      }
     });
     userService.setOnFailed(new EventHandler<WorkerStateEvent>() {
       @Override
@@ -87,25 +96,24 @@ public class AccountManagementController {
     });
   }
 
-  private ObservableList<UserData> parseUsers(){
+  private ObservableList<UserData> parseUsers() {
     List<String[]> userData = userService.getValue();
-    
+
     List<UserData> userList = new ArrayList<>();
 
-    for (int i = 0; i < userData.size(); i++){
-        String[] tempArray = userData.get(i);
-        UserData tempUser = new UserData(tempArray[0], tempArray[1], tempArray[2]);
-        userList.add(tempUser);
+    for (int i = 0; i < userData.size(); i++) {
+      String[] tempArray = userData.get(i);
+      UserData tempUser = new UserData(tempArray[0], tempArray[1], tempArray[2]);
+      userList.add(tempUser);
     }
 
-   
     return toObservableList(userList);
   }
 
-  private ObservableList<UserData> toObservableList(List<UserData> list){
+  private ObservableList<UserData> toObservableList(List<UserData> list) {
     ObservableList<UserData> observableList = FXCollections.observableArrayList();
 
-    for (int i = 0; i < list.size(); i++){
+    for (int i = 0; i < list.size(); i++) {
       observableList.add(list.get(i));
     }
